LogQL + PromQL: the stream and series identity becomes 128 bits (#498)
A stream or series was identified by a 64-bit hash with no comparison of
the underlying label set. Two label sets whose hash collided shared one
identity, and the loser's rows became unreachable rather than mislabelled:
a 200, no warning, no partial marker. Widening removes the defect rather
than detecting it; there is no collision detector.

The identity

  fp128(buf) = (cityHash64(buf) << 64) | xxHash64(buf, seed 0)

over the buffer each family already built. Both halves stay bit-identical
to a database function, so the value is still derivable server-side:
bitShiftLeft(toUInt128(cityHash64(buf)), 64) + toUInt128(xxHash64(buf)),
now pinned against a live server rather than resting on one hand-run
statement.

Fingerprint is a sealed newtype with no Display. Above 2^64 a bare decimal
literal is read as Float64, exact only to 2^53, so `= <bare>` silently
matches a neighbouring row and `IN (<bare list>)` prunes every granule.
The only spelling is toUInt128('<decimal>'), minted by
Fingerprint::sql_literal, and the wrong form is a compile error. Six
rendering sites take the minted literal; none of the five SQL modules has
a parameter typed as an identity.

The eight CREATEs carrying the column are edited in place at ids 4, 5, 6,
7, 8, 9, 23 and 29 — there is no tagged release and no persistent
deployment, which is the condition the amendment policy names. A developer
with an existing local database meets SchemaError::MigrationDrift and
drops it. No ALTER, no new id, no _dist or view template touched.

Two behaviours moved that the plan did not anticipate, both measured:
transform cannot RETURN a UInt128, so the group-key class widens outside
the call; and run_init issues eleven TTL/projection mutations on a fresh
database, so the criterion asserts no MODIFY COLUMN rather than an empty
mutation list.

+++ crates/pulsus-read/tests/live_discovery_fallback.rs
@@ -103,7 +103,7 @@ async fn init_db(bootstrap: &ChClient, db: &str) {
 #[derive(Row, serde::Serialize, serde::Deserialize, Debug, Clone)]
 struct SeedSeriesRow {
     metric_name: String,
-    fingerprint: u64,
+    fingerprint: u128,
     unix_milli: i64,
     labels: String,
 }
